Moved stat structs to stat.h and removed proc_stat from process_info
