+++ user/procstat.c
@@ -6,6 +6,7 @@ See the file LICENSE for details.
 
 #include "library/syscalls.h"
 #include "kernel/syscall.h"
+#include "kernel/stat.h"
 #include "library/string.h"
 
 int main(int argc, char const *argv[]) {
@@ -27,17 +28,17 @@ int main(int argc, char const *argv[]) {
 	struct process_info info;
 	process_wait(&info, -1);
 	unsigned int timeElapsed = gettimeofday() - startTime;
+	struct proc_stat stat;
+	process_stat(&stat, pid);
 	printf("Process %u exited with status %d\n", info.pid, info.exitcode);
 	printf("Time elapsed: %d:%d:%d\n", timeElapsed/3600, (timeElapsed%3600)/60, timeElapsed % 60);
-	printf("Disk 0: %d blocks read, %d blocks written\n", info.stat.blocks_read[0], info.stat.blocks_written[0]);
-	printf("Disk 1: %d blocks read, %d blocks written\n", info.stat.blocks_read[1], info.stat.blocks_written[1]);
-	printf("Disk 2: %d blocks read, %d blocks written\n", info.stat.blocks_read[2], info.stat.blocks_written[2]);
-	printf("Disk 3: %d blocks read, %d blocks written\n", info.stat.blocks_read[3], info.stat.blocks_written[3]);
+	printf("%d blocks read, %d blocks written\n", stat.blocks_read, stat.blocks_written);
+	printf("%d bytes read, %d bytes written\n", stat.bytes_read, stat.bytes_written);
 
 	printf("System calls used:\n");
 	for (int i = 0; i < MAX_SYSCALL; i++) {
-		if (info.stat.syscall_count[i]) {
-			printf("Syscall %d: %u\n", i, info.stat.syscall_count[i]);
+		if (stat.syscall_count[i]) {
+			printf("Syscall %d: %u\n", i, stat.syscall_count[i]);
 		}
 	}
